fix: transitive dependency on partialSpyOn renderer tests
