+++ src/apps/renderer/pages/Settings/General/sync-root-location.test.tsx
@@ -1,7 +1,7 @@
 import { renderHook } from '@testing-library/react-hooks';
 import * as useGetSyncRootLocation from '@/apps/renderer/api/use-get-sync-root-location';
 import { createAbsolutePath } from '@/context/local/localFile/infrastructure/AbsolutePath';
-import { partialSpyOn } from '@/tests/vitest/utils.helper.test';
+import { partialSpyOn } from '@/tests/vitest/partial-spy-on.helper';
 import { useSyncRootLocation } from './sync-root-location';
 
 describe('sync-root-location', () => {
